# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant Reservation API",
        "description": "API para gestionar reservas de restaurante",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/tables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get all tables",
                "operationId": "99272e10962b170fa754fc7b7e2ecdcf",
                "responses": {
                    "200": {
                        "description": "List of all tables",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Table"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Create a new table",
                "operationId": "3e9e6cbc406698544b85420629472d06",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero_mesa": {
                                        "type": "string",
                                        "example": "A-12"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "ubicacion": {
                                        "type": "string",
                                        "example": "Terraza Norte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Table created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tables/{id}": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get a table by ID",
                "operationId": "f08976f163aecc9693e20a370e0c2f6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the table",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Table not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tables"
                ],
                "summary": "Update a table",
                "operationId": "fec720a6666eddbcdbfc0c63f1129487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the table to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero_mesa": {
                                        "type": "string",
                                        "example": "A-14"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "ubicacion": {
                                        "type": "string",
                                        "example": "Terraza Sur"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Table updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Table not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tables"
                ],
                "summary": "Delete a table",
                "operationId": "ddbf89c289224ea4bee9db5dd7718fb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the table to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table deleted successfully"
                    },
                    "404": {
                        "description": "Table not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Table": {
                "title": "Table Model",
                "description": "Table model representing a restaurant table",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numero_mesa": {
                        "type": "string",
                        "example": "A-12"
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 4
                    },
                    "ubicacion": {
                        "type": "string",
                        "example": "Terraza Norte"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-15T19:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-15T19:30:00.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}